packet: reinforce the compatible version negotiation
Every time when version changes, crypto_initial_keys must be regenerated
and installed, so add quic_packet_version_change() helper to do this.

In quic_packet_handshake_header_process(), when processing Server Initial
packet, do a version change if the version in the packet is not the one
client choose but still a compatible version.

Signed-off-by: Xin Long <lucien.xin@gmail.com>