output: do not send packets with level in filter
Comparing application frames, Handshake frames (level > 0) should always
be sent first, even if it is in any pnspace's RTX timeout. So delete the
level from the filter to always send handshake frames in RTX timeout.

Also, filter in packet can be dropped, and we can use max_snd_count to
indicate if the filter is set in quic_packet_config().

Meanwhile, bring the sleep(1) back in perf_test.c, so in the netem test,
it will have time to retransmit the last msg if it gets lost.

Signed-off-by: Xin Long <lucien.xin@gmail.com>