pnspace: change QUIC_PN_MAX_GABS to 32
Change it to a smaller value, so that it can fit in function stack
and remove it from the struct quic_pnspace.

This not only reduces the size of quic_sock, but also improves the
performance when creating ACK frame if there are too many gap
blocks due to packet loss.

Signed-off-by: Xin Long <lucien.xin@gmail.com>