packet: use packet version during version negotiation
This patch is to make config version work as the preferred version and
only be used when choosing version in quic_packet_select_version() and
packet version will be used for the version it currently uses and it
may be updated when receiving INITIAL packets and selecting versions.

In quic_packet_select_version(), it changes to select version with
more factors considered, including local preference, remote chosen,
and support on both sides.

Meanwhile, fix some variable type casts in frame.c, and crypto resets
in quic_crypto_destroy().

Signed-off-by: Xin Long <lucien.xin@gmail.com>