output: initialize offset in quic_frame to -1 instead of 0
By initializing offset in quic_frame to -1 instead of 0, we no longer
need to do the extra check !pos->offset in quic_outq_transmitted_tail.
Also, with this chagne there is no need to reuse frame->skb to mark
if the frame->offset was set or not in quic_packet_pack_frames().

Signed-off-by: Xin Long <lucien.xin@gmail.com>