output: support stream_data append to last frame in stream_list
In quic_sendmsg(), before creating a stream data frame, we should first
check if the user data can be appended into last frame in stream_list.

As it is very likey the last frame is not fully the max size to the mss,
this will improve the sending efficiency and performance quite a lot.

Signed-off-by: Xin Long <lucien.xin@gmail.com>