socket: check hlist_empty before parsing alpn in initial packet
The crypto related modules are loaded when calling listen() or connect()
at the first time in process context. If there's never a socket created
in the sytem after the quic is loaded, when process an incoming request
it may try to load the crypto related modules in rx path (irq context)
in quic_packet_parse_alpn() and causes crash.

  [ ] Call Trace:
  [ ]  <TASK>
  [ ]  ttwu_do_activate+0x7f/0x170
  [ ]  try_to_wake_up+0x258/0x570
  [ ]  complete+0x41/0x80
  [ ]  process_one_work+0x1c7/0x380
  [ ]  worker_thread+0x4d/0x380
  [ ]  kthread+0xda/0x100
  [ ]  ret_from_fork+0x22/0x30
  [ ]  </TASK>

To avoid loading the crypto related modules in rx path (irq context),
we need to make sure there's already socket created in the hashtable
before parsing alpn in initial packet. Otherwise, there is no need
to continue the socket search with alpn.

Joint work with Moritz Buhl.

Signed-off-by: mbuhl <git@moritzbuhl.de>
Signed-off-by: Xin Long <lucien.xin@gmail.com>