cong: introduce min_window to cache the minimum cong window
Introduce min_window in quic_cong, and update it whenever the mss is
changed by:

  max(min(mss * 10, 14720U), mss * 2)

min_window will replace 'mss * 2' in congestion control algorithms,
so that the window will not becomes too small in beginning of slow
start. Update the unit tests accordingly.

While at it, add U at the end of QUIC_RTT_INIT/MIN/MAX to fix the
compiling warning for the clamp() call in quic_cong_pto_update(),
and use recvmsg to replace msleep in sample_test.

Signed-off-by: Xin Long <lucien.xin@gmail.com>