packet: check udp payload length when processing an initial packet
As Billy Foster noticed, rfc9000#section-14.1 provides: "A server MUST
discard an Initial packet that is carried in a UDP datagram with a
payload that is smaller than the smallest allowed maximum datagram size
of 1200 bytes."

This patch is to add the missing check for payload length of UDP packet
carrying initial packets in quic_packet_handshake_header_process() on
server side.

Meanwhile, in quic_packet_handshake_create() add padding to all initial
packets on both client and server sides.

Signed-off-by: Xin Long <lucien.xin@gmail.com>