protocol: separate protocol family ops from address family ops
As IPv6 socket might support binding to IPv4 addresses, so we need two
kind of family ops: one for address operations and the other one for
socket ops.

This patch is to separate protocol family ops from address family ops.
This enables IPv6 socket binding to IPv4 addresses and prepares for
the connection migration crossing IPv4 and IPv6 in the next patch.

Signed-off-by: Xin Long <lucien.xin@gmail.com>