packet: do not send retire_conn_id if received dcid is active
According to rfc9000#section-9.5:

  Due to network changes outside the control of its peer, an endpoint
  might receive packets from a new source address with the same
  Destination Connection ID field value, in which case it MAY continue
  to use the current connection ID with the new remote address while
  still sending from the same local address.

So we should not change the alt dcid to another one if the dcid in the
packet is the current active one.

Meanwhile, it also adds conn_id in quic_crypto_cb to save the received
conn_id so that it can update the source active conn_id every time
after path change in quic_outq_free_path().

Signed-off-by: Xin Long <lucien.xin@gmail.com>