cong: initialize the smoothed_rtt with the first valid ack
smoothed_rtt is initialized in quic_cong_set_config() by the default
value 333000 or by users. It must be update/set (not just calculated)
by the 1st SACK from the peer. Otherwise, it may take long time for
the calculation to reduce the impact by initialized smoothed_rtt.

Meanwhile, change QUIC_RTO_MIN to 30000 and max pto_count to 3 to
improve the performance on packet loss and retransmission.

Signed-off-by: Xin Long <lucien.xin@gmail.com>