packet: set errcode to ERROR_CRYPTO on handshake decryption failure
Currently, if a connection is closed but the CLOSE frame is lost, and a
new connection attempt is made with the same 4-tuple, the initial packet
is silently discarded due to decryption failure until the connection
times out.

This patch sets errcode to ERROR_CRYPTO when handshake decryption fails,
ensuring the connection is closed immediately. This allows a new
connection request with the same 4-tuple to be accepted as soon
as possible.

Signed-off-by: Xin Long <lucien.xin@gmail.com>