packet: optimize sk existence check by reusing time in space
Set the time_in_space field when any request socket is accepted in
quic_accept().

Modify quic_accept_sock_exists() to perform quic_sock_lookup() only
if the time in the skb control block is later than time_in_space.

This approach improves reliability compared to checking the backlog
in the skb control block, which fails in the following scenario:

   CPU1                         CPU2

   quic_packet_get_sock()
                                quic_accept()
   quic_packet_listen_process()

Signed-off-by: Xin Long <lucien.xin@gmail.com>