libquic: move SEND/RECV logging out of quic_handshake_next_step
In quic_handshake_next_step(), the log for Handshake SEND and RECV
does not print level, and the rmsg->iov.iov_len printed is always
sizeof(rmsg->data) (65536).

Thus add a level field to the internal quic_smsg and quic_rmsg
structs and include it in the log output, and move the logging
of SEND/RECV to quic_handshake_send/recv_msg_process(). It also
changes to use 'SENT' for better wording.

Additionally, set rmsg->level in quic_check_rmsg_level() and make
it return 0 upon a successful check.

Signed-off-by: Xin Long <lucien.xin@gmail.com>