socket: implement SIOCINQ, SIOCOUTQ and SIOCOUTQNSD ioctls
This patch adds support for the ioctl API on QUIC sockets, implementing
the following commands:

  SIOCINQ: returns the number of bytes in the receive queue.

  SIOCOUTQ: returns the number of bytes in the send queue.

  SIOCOUTQNSD: returns the number of unsent bytes.

To support SIOCOUTQNSD, a new unsent_bytes field is added to struct
quic_outqueue. This counter is incremented when user data is enqueued
in the send queue and decremented when bytes are actually sent.

Once the handshake is complete and the application uses only a single
stream for sending and receiving, this ioctl interface behaves similarly
to TCP.

Additionally, quic_outq_retransmit_list() no longer frees frames, as it
is called only when sending has failed. These frames (including dgram
and ping frames) should be retransmitted.

Signed-off-by: Xin Long <lucien.xin@gmail.com>