uapi: replace literal flag values with existing MSG_ constants
Replace hardcoded stream and datagram flag values in enum quic_msg_flags
with their equivalent MSG_* constants defined in the kernel.

This change also helps reduce the chance of conflicting or misaligned
values as the QUIC socket API evolves.

As 0x20000 is not defined, and 0x10 is defined differently between kernel
and userspace, change to use MSG_ERRQUEUE for MSG_STREAM_SNDBLOCK and
MSG_RST for MSG_DATAGRAM.

Suggested-by: Stefan Metzmacher <metze@samba.org>
Signed-off-by: Xin Long <lucien.xin@gmail.com>