crypto: decouple tfm allocation from secret installation
This change separates the allocation of the tfm (transform) from the
setting of the secret, allowing the transform to be reused across key
updates.

Previously, quic_packet_version_change() had to free and reallocate the
tfm when updating the secret. With this refactor, the tfm is allocated
once via quic_crypto_set_cipher(), and then reused when installing new
secrets.

This also enables the use of crypto->secret_tfm for hkdf_expand() during
initial key installation in connect(), accept(), and listen() paths,
improving efficiency and simplifying the code.

Signed-off-by: Xin Long <lucien.xin@gmail.com>