uapi: replace uintN_t and intN_t with __uN and __sN
This patch is to replace uintN_t and intN_t with __uN and __sN and
replace __uapi_quic_h__  and __linux_quic_h__ with _UAPI_LINUX_QUIC_H
and _LINUX_QUIC_H, as David Howells suggested.

Meanwhile, delete the unnecessary __net_quic_h__ in socket.h, and
assign values to QUIC_TRANSPORT_ERROR_*, which are defined in
rfc9000#section-20.1.

Signed-off-by: Xin Long <lucien.xin@gmail.com>