socket: avoid unhash when changing sk state to CLOSED
Syzkaller can still trigger a crash:

  RIP: 0010:quic_frame_stop_sending_create+0x20c/0x260 net/quic/frame.c:642
  Call Trace:
   <TASK>
   quic_frame_create+0x88/0x2d0 net/quic/frame.c:2185
   quic_outq_transmit_frame+0x32/0xc0 net/quic/outqueue.c:1165
   quic_sock_stream_stop_sending net/quic/socket.c:1518 [inline]
   quic_do_setsockopt+0x6e7/0x3890 net/quic/socket.c:2010
   quic_setsockopt+0x98/0xa0 net/quic/socket.c:2060
   do_sock_setsockopt+0xf8/0x1e0 net/socket.c:2344
   __sys_setsockopt+0x11f/0x1a0 net/socket.c:2369

Normally, connect() fails if the socket is still hashed. However, because
the code currently performs an unhash when transitioning the socket state
to CLOSED, it is possible to turn a listening socket into a connecting
socket with the sequence:

  listen() -> shutdown() -> connect()

This leads to invalid states and the reported crash.

Fix this by not performing unhash when changing the socket state to CLOSED.

In addition, ensure that exiting the listen state when backlog is zero
only applies to listening sockets. Without this, it is possible to turn
an accepted socket into a connecting socket via:

  accept() -> listen(backlog:0) -> connect()

which can also lead to this crash.

Signed-off-by: Xin Long <lucien.xin@gmail.com>