frame: reserve max varint length when creating stream frames
Syzkaller reported an out-of-bound crash:

  BUG: KASAN: slab-out-of-bounds in quic_put_var+0x183/0x190
    net/quic/common.c:221
  Call Trace:
   <TASK>
   __dump_stack lib/dump_stack.c:94 [inline]
   dump_stack_lvl+0x116/0x1b0 lib/dump_stack.c:120
   print_address_description mm/kasan/report.c:378 [inline]
   print_report+0xca/0x5f0 mm/kasan/report.c:482
   kasan_report+0xca/0x100 mm/kasan/report.c:595
   quic_put_var net/quic/common.c:237 [inline]
   quic_put_var+0x183/0x190 net/quic/common.c:221
   quic_frame_stream_append+0x4f8/0x900 net/quic/frame.c:2318
   quic_outq_stream_append+0x1aa/0x4e0 net/quic/outqueue.c:353
   quic_sendmsg+0x1148/0x1ee0 net/quic/socket.c:989

The crash occurs when appending user data after the MSS(max_frame_len)
changed to a big value. The allocated frame->data buffer is too small
to hold the updated stream data length, leading to the out-of-bounds
write.

Fix this by always reserving space for the largest possible varint
length when creating stream frames.

Signed-off-by: Xin Long <lucien.xin@gmail.com>