packet: introduce backlog work for key change on rx path
Syzkaller reported a warning on rx path:

  BUG: sleeping function called from invalid context at
    ./include/linux/sched/mm.h:321
  in_atomic(): 1, irqs_disabled(): 0, non_block: 0, pid: 12693,
    name: syz.5.22
  Call Trace:
   <IRQ>
   dump_stack_lvl+0x180/0x1b0
   __might_resched+0x359/0x580
   __kmalloc_noprof+0x383/0x540
   crypto_aead_setkey+0xc8/0x2c0
   quic_crypto_tx_keys_derive_and_install+0x36b/0x4a0
   quic_crypto_set_secret+0x12c/0x410
   quic_crypto_initial_keys_install+0x347/0x4b0
   quic_packet_parse_alpn+0x54a/0x10d0
   quic_packet_rcv+0x1011/0x14d0
   quic_udp_rcv+0x44c/0x580
   udpv6_queue_rcv_one_skb+0xbb6/0x1a70
   udpv6_queue_rcv_skb+0x192/0xd10
   udp6_unicast_rcv_skb+0x158/0x300
   __udp6_lib_rcv+0x1e33/0x30a0
   ip6_protocol_deliver_rcu+0x6f5/0x1410
   ip6_input_finish+0x1ba/0x4c0
   ip6_input+0x10b/0x300
   ipv6_rcv+0x267/0x650

This happens because key changes on the RX path may sleep and therefore
must run in process context.

The affected operations include:

  - quic_packet_parse_alpn()
  - quic_packet_retry_create()
  - quic_crypto_verify_token()
  - quic_packet_refuse_close_create()
  - quic_packet_version_process()
  - quic_packet_retry_process()
  - quic_crypto_key_update()

These are not part of the critical data path, so this patch introduces a
per-namespace backlog work item. When a packet that requires key changes
is received outside process context, it is enqueued into the namespace’s
backlog_list and processed later by the backlog work.

Additionally, cb->backlog is set for sk_add_backlog(), which ensures the
packet is handled in process context as well.

Signed-off-by: Xin Long <lucien.xin@gmail.com>