socket: call sk_common_release() on free path of quic_accept()
quic_close() sends a CLOSE frame, sets the socket state to CLOSE,
and then calls sk_common_release().

On the free path of quic_accept(), the socket state is already
CLOSE, so there's no need to send a CLOSE frame again. In this
case, calling sk_common_release() directly is sufficient to free
the socket, and SINGLE_DEPTH_NESTING is no longer required in
quic_close().

Signed-off-by: Xin Long <lucien.xin@gmail.com>