protocol: add a shared workqueue for all QUIC works
Introduce a dedicated workqueue shared by all QUIC components, replacing
direct schedule_work() calls with queue_work(quic_wq, ...). This allows
better control of asynchronous tasks and ensures proper synchronization
and cleanup during module exit.

This is needed because some works, such as us->work and qn->work, do not
hold a reference to the QUIC socket or module. We must ensure that all
pending works are completed before the QUIC module is unloaded.

The workqueue is created in quic_init() and destroyed in quic_exit().

Signed-off-by: Xin Long <lucien.xin@gmail.com>